@@ -7,8 +7,8 @@ import android.content.pm.PackageManager;
 import android.net.Uri;
 import android.os.Build;
 import android.os.Environment;
-import android.support.v4.app.ActivityCompat;
-import android.support.v4.content.ContextCompat;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.webkit.MimeTypeMap;
 
@@ -141,8 +141,8 @@ public class FilePickerPlugin implements MethodCallHandler {
 
     if(isCustom) {
       final String extension = type.split("__CUSTOM_")[1].toLowerCase();
-//      String mime = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
-//      Log.i(TAG, "Custom file type: " + mime);
+      String mime = MimeTypeMap.getSingleton().getMimeTypeFromExtension(extension);
+      Log.i(TAG, "Custom file type: " + mime);
       return "audio/*";
     }
 
